## Related Pull Requests & Issues

- #699
- MTES-MCT/monitorfish#2373

----

- [x] Tests E2E (Cypress)

+++ backend/src/main/kotlin/fr/gouv/cacem/monitorenv/domain/use_cases/controlUnit/DeleteControlUnit.kt
@@ -2,10 +2,20 @@ package fr.gouv.cacem.monitorenv.domain.use_cases.controlUnit
 
 import fr.gouv.cacem.monitorenv.config.UseCase
 import fr.gouv.cacem.monitorenv.domain.repositories.IControlUnitRepository
+import fr.gouv.cacem.monitorenv.infrastructure.database.repositories.exceptions.ForeignKeyConstraintException
 
 @UseCase
-class DeleteControlUnit(private val controlUnitRepository: IControlUnitRepository) {
+class DeleteControlUnit(
+    private val controlUnitRepository: IControlUnitRepository,
+    private val canDeleteControlUnit: CanDeleteControlUnit,
+) {
     fun execute(controlUnitId: Int) {
+        if (!canDeleteControlUnit.execute(controlUnitId)) {
+            throw ForeignKeyConstraintException(
+                "Cannot delete control unit  (ID=$controlUnitId) due to existing relationships.",
+            )
+        }
+
         return controlUnitRepository.deleteById(controlUnitId)
     }
 }
